fix(video.controller.ts): Fixed error when uploading non-video files
Removed the 'Bad Request' exception thrown when the uploaded file's MIME type did not start with 'video/', as it was preventing users from uploading non-video files. Instead, a new variable was added to create a new file name using a randomly generated ID and the extension of the uploaded file.